+++ src/services/finders/CardFinder.java
@@ -15,14 +15,14 @@ import services.DatabaseManager;
 public final class CardFinder {
 	private static final String tableName = "Cards";
 	
-	public static ResultSet find(int cardID) {
+	public static ResultSet find(long cardID) {
 		String query = "SELECT * FROM " + tableName + " WHERE id = ?;";
 		Connection conn = DatabaseManager.getConnection();
 		ResultSet output = null;
 		
 		try {
 			PreparedStatement ps = conn.prepareStatement(query);
-			ps.setInt(1, cardID);
+			ps.setLong(1, cardID);
 			
 			output = ps.executeQuery();
 			
@@ -42,14 +42,14 @@ public final class CardFinder {
 		return output;
 	}
 	
-	public static ResultSet findCardByDeckPosition(int deckID, int deckPosition) {
+	public static ResultSet findCardByDeckPosition(long deckID, int deckPosition) {
 		String query = "SELECT * FROM " + tableName + " WHERE deck_id = ? AND deck_position = ?;";
 		Connection conn = DatabaseManager.getConnection();
 		ResultSet output = null;
 		
 		try {
 			PreparedStatement ps = conn.prepareStatement(query);
-			ps.setInt(1, deckID);
+			ps.setLong(1, deckID);
 			ps.setInt(2, deckPosition);
 			
 			output = ps.executeQuery();
@@ -70,14 +70,14 @@ public final class CardFinder {
 		return output;
 	}
 	
-	public static ResultSet findCardsForUserID(int userID) {
+	public static ResultSet findCardsForUserID(long userID) {
 		String query = "SELECT Cards.* FROM Users JOIN Decks ON Users.id = Decks.user_id JOIN Cards ON Decks.id = Cards.deck_id WHERE Decks.user_id = ?;";
 		Connection conn = DatabaseManager.getConnection();
 		ResultSet output = null;
 		
 		try {
 			PreparedStatement ps = conn.prepareStatement(query);
-			ps.setInt(1, userID);
+			ps.setLong(1, userID);
 			
 			output = ps.executeQuery();
