+++ src/services/finders/DeckFinder.java
@@ -15,14 +15,14 @@ import services.DatabaseManager;
 public class DeckFinder {
 	private static final String tableName = "Decks";
 	
-	public static ResultSet find(int id) {
+	public static ResultSet find(long id) {
 		String query = "SELECT * FROM " + tableName + " WHERE id = ?;";
 		Connection conn = DatabaseManager.getConnection();
 		ResultSet output = null;
 		
 		try {
 			PreparedStatement ps = conn.prepareStatement(query);
-			ps.setInt(1, id);
+			ps.setLong(1, id);
 			
 			output = ps.executeQuery();
 			
@@ -42,14 +42,14 @@ public class DeckFinder {
 		return output;
 	}
 	
-	public static ResultSet findByUserID(int userID) {
+	public static ResultSet findByUserID(long userID) {
 		String query = "SELECT * FROM " + tableName + " WHERE user_id = ?;";
 		Connection conn = DatabaseManager.getConnection();
 		ResultSet output = null;
 		
 		try {
 			PreparedStatement ps = conn.prepareStatement(query);
-			ps.setInt(1, userID);
+			ps.setLong(1, userID);
 			
 			output = ps.executeQuery();
