+++ src/services/finders/UserFinder.java
@@ -16,14 +16,14 @@ import services.UserHelper;
 public class UserFinder {
 	private static final String tableName = "Users";
 	
-	public static ResultSet find(int id) {
+	public static ResultSet find(long id) {
 		String query = "SELECT * FROM " + tableName + " WHERE id = ?;";
 		Connection conn = DatabaseManager.getConnection();
 		ResultSet output = null;
 		
 		try {
 			PreparedStatement ps = conn.prepareStatement(query);
-			ps.setInt(1, id);
+			ps.setLong(1, id);
 			
 			output = ps.executeQuery();
